backend/requirements.txt
# FEAST Backend - FastAPI

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0

# Fuzzy Matching
rapidfuzz>=3.5.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# CORS
# (included in fastapi)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Production server
gunicorn>=21.0.0
